refactor: component created to better maintenance
